requirements.txt
Django==3.1.2
pandas==1.1.2
django_heroku==0.3.1
drf_extra_fields==3.0.2
whitenoise==5.2.0
requests==2.24.0
knox==0.1.0
gunicorn==20.0.4
numpy==1.19.2
djangorestframework==3.12.1
